MGDOBR-1214: Move some more bits to -core
