+++ shard-operator-parent/shard-operator-core/src/main/java/com/redhat/service/smartevents/shard/operator/core/providers/GlobalConfigurationsProviderImpl.java
@@ -1,4 +1,4 @@
-package com.redhat.service.smartevents.shard.operator.v1.providers;
+package com.redhat.service.smartevents.shard.operator.core.providers;
 
 import javax.enterprise.context.ApplicationScoped;
