+++ shard-operator-parent/shard-operator-v1/src/main/java/com/redhat/service/smartevents/shard/operator/v1/networking/NetworkingServiceProducer.java
@@ -6,7 +6,7 @@ import javax.inject.Singleton;
 
 import com.redhat.service.smartevents.infra.core.app.Orchestrator;
 import com.redhat.service.smartevents.infra.core.app.OrchestratorConfigProvider;
-import com.redhat.service.smartevents.shard.operator.v1.providers.IstioGatewayProvider;
+import com.redhat.service.smartevents.shard.operator.core.providers.IstioGatewayProvider;
 import com.redhat.service.smartevents.shard.operator.v1.providers.TemplateProvider;
 
 import io.fabric8.openshift.client.OpenShiftClient;
