+++ shard-operator-parent/shard-operator-v1/src/main/java/com/redhat/service/smartevents/shard/operator/v1/resources/BridgeExecutorStatus.java
@@ -2,10 +2,17 @@ package com.redhat.service.smartevents.shard.operator.v1.resources;
 
 import java.util.HashSet;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.redhat.service.smartevents.infra.core.models.ManagedResourceStatus;
 import com.redhat.service.smartevents.shard.operator.core.resources.Condition;
+import com.redhat.service.smartevents.shard.operator.core.resources.ConditionReasonConstants;
 import com.redhat.service.smartevents.shard.operator.core.resources.ConditionStatus;
 import com.redhat.service.smartevents.shard.operator.core.resources.ConditionTypeConstants;
+import com.redhat.service.smartevents.shard.operator.core.resources.CustomResourceStatus;
+import com.redhat.service.smartevents.shard.operator.core.utils.DeploymentStatusUtils;
+
+import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.client.internal.readiness.Readiness;
 
 public class BridgeExecutorStatus extends CustomResourceStatus {
 
@@ -44,6 +51,58 @@ public class BridgeExecutorStatus extends CustomResourceStatus {
             return ManagedResourceStatus.FAILED;
         }
         return ManagedResourceStatus.PROVISIONING;
+    }
+
+    @JsonIgnore
+    public final void setStatusFromDeployment(Deployment deployment) {
+        if (deployment.getStatus() == null) {
+            if (!isConditionTypeFalse(ConditionTypeConstants.READY, ConditionReasonConstants.DEPLOYMENT_NOT_AVAILABLE)) {
+                markConditionFalse(ConditionTypeConstants.READY,
+                        ConditionReasonConstants.DEPLOYMENT_NOT_AVAILABLE,
+                        "");
+            }
+            if (!isConditionTypeFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE)) {
+                markConditionFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE);
+            }
+        } else if (Readiness.isDeploymentReady(deployment)) {
+            if (!isConditionTypeFalse(ConditionTypeConstants.READY, ConditionReasonConstants.DEPLOYMENT_AVAILABLE)) {
+                markConditionFalse(ConditionTypeConstants.READY,
+                        ConditionReasonConstants.DEPLOYMENT_AVAILABLE,
+                        "");
+            }
+            if (!isConditionTypeTrue(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE)) {
+                markConditionTrue(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE);
+            }
+        } else {
+            if (DeploymentStatusUtils.isTimeoutFailure(deployment)) {
+                if (!isConditionTypeFalse(ConditionTypeConstants.READY, ConditionReasonConstants.DEPLOYMENT_FAILED)) {
+                    markConditionFalse(ConditionTypeConstants.READY,
+                            ConditionReasonConstants.DEPLOYMENT_FAILED,
+                            DeploymentStatusUtils.getReasonAndMessageForTimeoutFailure(deployment));
+                }
+                if (!isConditionTypeFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE)) {
+                    markConditionFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE);
+                }
+            } else if (DeploymentStatusUtils.isStatusReplicaFailure(deployment)) {
+                if (!isConditionTypeFalse(ConditionTypeConstants.READY, ConditionReasonConstants.DEPLOYMENT_FAILED)) {
+                    markConditionFalse(ConditionTypeConstants.READY,
+                            ConditionReasonConstants.DEPLOYMENT_FAILED,
+                            DeploymentStatusUtils.getReasonAndMessageForReplicaFailure(deployment));
+                }
+                if (!isConditionTypeFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE)) {
+                    markConditionFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE);
+                }
+            } else {
+                if (!isConditionTypeFalse(ConditionTypeConstants.READY, ConditionReasonConstants.DEPLOYMENT_NOT_AVAILABLE)) {
+                    markConditionFalse(ConditionTypeConstants.READY,
+                            ConditionReasonConstants.DEPLOYMENT_NOT_AVAILABLE,
+                            "");
+                }
+                if (!isConditionTypeFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE)) {
+                    markConditionFalse(BridgeExecutorStatus.DEPLOYMENT_AVAILABLE);
+                }
+            }
+        }
     }
 
 }
