+++ shard-operator-parent/shard-operator-v1/src/test/java/com/redhat/service/smartevents/shard/operator/v1/providers/IstioGatewayProviderMock.java
@@ -1,5 +1,7 @@
 package com.redhat.service.smartevents.shard.operator.v1.providers;
 
+import com.redhat.service.smartevents.shard.operator.core.providers.IstioGatewayProviderImpl;
+
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
 import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServiceBuilder;
